requirements.txt
pytest==8.4.2
PyYAML==6.0.3
selenium==4.36.0
